requirements.txt
streamlit
langchain-mistralai
langchain-chroma
python-dotenv
chromadb